backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
python-multipart>=0.0.9
httpx>=0.28
